+++ assets/js/footer.js
@@ -13,6 +13,7 @@ footer.innerHTML = `
     <li><p><a href="tts-docs.html">TTS Documentation</a></p></li>
     <li><p><a href="demos.html">TTS Voice Demos</a></p></li>
     <li><p><a href="credits.html">Credits</a></p></li>
+    <li><p><a href="changelog.html">Changelog</a></p></li>
 </ul>
 <br class="mobileBreak">
 <div id="footer-mobile-arrow" style="display: none;"></div>